# go-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A comprehensive anonymous form builder and response collection system",
        "title": "AnoQ Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "AnoQ Support",
            "email": "support@anoq.com"
        },
        "license": {
            "name": "MIT",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/response": {
            "post": {
                "description": "Submit answers to a form (public endpoint)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Responses"
                ],
                "summary": "Submit a form response",
                "parameters": [
                    {
                        "description": "Form response data",
                        "name": "response",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateResponseRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Response submitted successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "response_id": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body or form not accepting responses",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Form not found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/response/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get details of a specific form response",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Responses"
                ],
                "summary": "Get a response by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Response ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response details",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "response": {
                                    "$ref": "#/definitions/model.ResponseDetailResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid response ID",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Response not found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.AnswerResponse": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "question": {
                    "$ref": "#/definitions/model.QuestionResponse"
                },
                "question_id": {
                    "type": "string"
                },
                "selected_choices": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "model.CreateAnswerRequest": {
            "description": "Request payload for submitting an answer to a question",
            "type": "object",
            "required": [
                "question_id"
            ],
            "properties": {
                "answer": {
                    "description": "Answer text for basic questions",
                    "type": "string",
                    "example": "Very satisfied"
                },
                "question_id": {
                    "description": "Question ID being answered (required)",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440003"
                },
                "selected_choices": {
                    "description": "Selected choices for multiple_choice questions",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"Very satisfied\"]"
                    ]
                }
            }
        },
...
        "model.CreateResponseRequest": {
            "description": "Request payload for submitting a form response",
            "type": "object",
            "required": [
                "answers",
                "form_id"
            ],
            "properties": {
                "answers": {
                    "description": "List of answers to form questions (required)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CreateAnswerRequest"
                    }
                },
                "email": {
                    "description": "Optional respondent email",
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "form_id": {
                    "description": "Form ID to submit response for (required)",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440002"
                },
                "name": {
                    "description": "Optional respondent name",
                    "type": "string",
                    "example": "John Doe"
                }
            }
        },
...
        "model.FormResponse": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/model.UserResponse"
                },
                "author_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "modified_at": {
                    "type": "string"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.QuestionResponse"
                    }
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.FormStatus"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "model.FormStatus": {
            "type": "string",
            "enum": [
                "open",
                "closed"
            ],
            "x-enum-varnames": [
                "FormStatusOpen",
                "FormStatusClosed"
            ]
        },
...
        "model.QuestionResponse": {
            "type": "object",
            "properties": {
                "allow_multiple": {
                    "type": "boolean"
                },
                "answer": {
                    "type": "string"
                },
                "choices": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "form_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "question_text": {
                    "type": "string"
                },
                "required": {
                    "type": "boolean"
                },
                "selected_choice": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "$ref": "#/definitions/model.QuestionType"
                }
            }
        },
        "model.QuestionType": {
            "type": "string",
            "enum": [
                "basic",
                "multiple_choice"
            ],
            "x-enum-varnames": [
                "QuestionTypeBasic",
                "QuestionTypeMultipleChoice"
            ]
        },
        "model.ResponseDetailResponse": {
            "type": "object",
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AnswerResponse"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "form": {
                    "$ref": "#/definitions/model.FormResponse"
                },
                "form_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "modified_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "user_ip": {
                    "type": "string"
                }
            }
        },
...
        "model.UserResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "family_name": {
                    "type": "string"
                },
                "given_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "SessionToken": {
            "description": "Session token for authentication",
            "type": "apiKey",
            "name": "X-Session-Token",
            "in": "header"
        }
    }
}